+++ src/main/java/nl/dtls/fairdatapoint/api/controller/exception/ExceptionControllerAdvice.java
@@ -35,8 +35,11 @@ import nl.dtls.fairdatapoint.api.dto.error.ErrorDTO;
 import nl.dtls.fairdatapoint.entity.exception.*;
 import nl.dtls.fairdatapoint.entity.index.exception.IndexException;
 import nl.dtls.fairdatapoint.service.metadata.exception.MetadataServiceException;
+import org.eclipse.rdf4j.common.exception.RDF4JException;
 import org.eclipse.rdf4j.model.IRI;
 import org.eclipse.rdf4j.model.Model;
+import org.eclipse.rdf4j.query.MalformedQueryException;
+import org.eclipse.rdf4j.query.QueryEvaluationException;
 import org.eclipse.rdf4j.rio.RDFFormat;
 import org.eclipse.rdf4j.rio.Rio;
 import org.springframework.http.HttpStatus;
@@ -50,6 +53,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 import java.io.StringWriter;
+import java.util.Map;
 
 import static java.lang.String.format;
 import static nl.dtls.fairdatapoint.util.RdfIOUtil.getWriterConfig;
@@ -104,6 +108,44 @@ public class ExceptionControllerAdvice {
         return validationReportModel;
     }
 
+    @ExceptionHandler(MalformedQueryException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ResponseBody
+    @ApiResponse(
+            responseCode = "400",
+            description = "Bad request",
+            content = @Content(
+                    mediaType = MediaType.APPLICATION_JSON_VALUE,
+                    schema = @Schema(implementation = ErrorDTO.class)
+            )
+    )
+    public ErrorDTO handleInvalidQuery(MalformedQueryException e) {
+        String message = "Invalid SPARQL query";
+        log.error(message);
+        log.debug("Handling invalid query (MalformedQueryException)", e);
+        Map<String, String> details = Map.of("sparql", e.getMessage(), "exception", e.getClass().getName());
+        return new ErrorDTO(HttpStatus.BAD_REQUEST, "Invalid SPARQL query", details);
+    }
+
+    @ExceptionHandler(QueryEvaluationException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ResponseBody
+    @ApiResponse(
+            responseCode = "400",
+            description = "Bad request",
+            content = @Content(
+                    mediaType = MediaType.APPLICATION_JSON_VALUE,
+                    schema = @Schema(implementation = ErrorDTO.class)
+            )
+    )
+    public ErrorDTO handleInvalidQuery(QueryEvaluationException e) {
+        String message = "Invalid SPARQL query";
+        log.error(message);
+        log.debug("Handling invalid query (QueryEvaluationException)", e);
+        Map<String, String> details = Map.of("sparql", e.getMessage(), "exception", e.getClass().getName());
+        return new ErrorDTO(HttpStatus.BAD_REQUEST, "Invalid SPARQL query", details);
+    }
+
     @ExceptionHandler({BadCredentialsException.class, UnauthorizedException.class})
     @ResponseStatus(HttpStatus.UNAUTHORIZED)
     @ResponseBody
@@ -117,7 +159,7 @@ public class ExceptionControllerAdvice {
     )
     public ErrorDTO handleUnauthorized(Exception e) {
         log.error(e.getMessage());
-        e.printStackTrace();
+        log.debug("Handling unauthorized", e);
         return new ErrorDTO(HttpStatus.UNAUTHORIZED, e.getMessage());
     }